+++ app/src/main/res/layout/fragment_dictionary.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".DictionaryFragment">
+   >
 
     <androidx.appcompat.widget.SearchView
         android:id="@+id/searchBar"
