+++ app/src/main/res/layout/fragment_training.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:background="@color/primaryOrange"
     android:layout_height="match_parent"
-    tools:context=".TrainingFragment">
+   >
 
 
     <TextView
