+++ app/src/main/res/layout/fragment_video.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:background="@color/inkGrey"
     android:layout_height="match_parent"
-    tools:context=".VideoFragment">
+   >
 
     <TextView
         android:layout_width="match_parent"
