+++ app/src/main/res/menu/bottom_menu.xml
@@ -4,13 +4,13 @@
     <item
         android:id="@+id/dictionaryFragment"
         android:icon="@drawable/dictionary_ic"
-        android:title="Dictionary" />
+        android:title="@string/dictionary" />
     <item
         android:id="@+id/trainingFragment"
         android:icon="@drawable/training_ic"
-        android:title="Training" />
+        android:title="@string/training" />
     <item
         android:id="@+id/videoFragment"
         android:icon="@drawable/video_ic"
-        android:title="Video" />
+        android:title="@string/video" />
 </menu>
